Use hypermerge crawler
