+++ src/renderer/components/content-types/workspace/Workspace.tsx
@@ -11,7 +11,7 @@ import { ContactDoc } from '../contact'
 import * as WebStreamLogic from '../../../../WebStreamLogic'
 
 import './Workspace.css'
-import { useDocument } from '../../../Hooks'
+import { useDocument, useCrawler } from '../../../Hooks'
 import {
   useAllHeartbeats,
   useHeartbeat,
@@ -25,7 +25,6 @@ import { CurrentDeviceContext } from './Device'
 import WorkspaceInList from './WorkspaceInList'
 import { importPlainText } from '../../../ImportData'
 import * as DataUrl from '../../../../DataUrl'
-import { usePreload } from '../../../Preload'
 
 const log = Debug('pushpin:workspace')
 
@@ -52,7 +51,7 @@ export default function Workspace(props: WorkspaceContentProps) {
   const [workspace, changeWorkspace] = useDocument<Doc>(props.hypermergeUrl)
   const currentDeviceUrl = useContext(CurrentDeviceContext)
 
-  usePreload(props.hypermergeUrl)
+  useCrawler(props.hypermergeUrl)
 
   const selfId = workspace && workspace.selfId
   const currentDocUrl =
